socialproject/requirements.txt
asgiref==3.11.1
celery==5.3.4
django==6.0.3
django-cors-headers==4.9.0
django-filter==25.1
django-redis==5.4.0
django-stubs==6.0.1
django-stubs-ext==6.0.1
djangorestframework==3.17.0
djangorestframework-simplejwt==5.5.1
djangorestframework-stubs==3.16.8
pyjwt==2.12.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
sqlparse==0.5.5
types-pyyaml==6.0.12.20250915
typing-extensions==4.15.0
tzdata==2025.3
